Refactor changeset.py into a module
